Former-commit-id: 298fda7cf8e070684e4b4dab749803b58663ecb1 [formerly b94993e7fcd92cf35c5b864d81e39f78b2e0ccca]
Former-commit-id: c5acb6b700b373c3be3d882f7deb1a6f862a84f2

+++ av/NativeAudioNode.h
@@ -207,9 +207,9 @@ class NativeAudioNodeDelay : public NativeAudioNode
             DELAY, WET, DRY 
         };
 
-        double delay;
-        int wet;
-        int dry;
+        int delay;
+        double wet;
+        double dry;
         float **buffers;
 
         virtual bool process();
